Leave-one-study or leave-one-effect sensitivity analysis

Development-snapshot manual for apm_leave_one_out. The roxygen source in R/ is authoritative; regenerate with roxygen2 during local validation.

Usage

apm_leave_one_out(...)

Arguments

...

See the complete roxygen documentation and vignettes for arguments and method-specific assumptions.

Value

An auditable agriPairMetaFlow result or output object as documented in the function source.

Examples

# Example 1: independent yield-response studies.
apm_leave_one_out(apm_fit(agri_effects_benchmark), unit="study")
#> <apm_sensitivity>
#>  omitted estimate         se ci_lower ci_upper  pi_lower pi_upper
#>     BM01 1.090122 0.02672415 1.090104 1.090140 0.9248696 1.284901
#>     BM02 1.088632 0.02629307 1.088614 1.088650 0.9288865 1.275850
#>     BM03 1.091326 0.02674515 1.091308 1.091345 0.9243547 1.288459
#>     BM04 1.095947 0.02742421 1.095928 1.095966 0.9211461 1.303918
#>     BM05 1.095182 0.02728954 1.095163 1.095200 0.9215091 1.301585
#>     BM06 1.086627 0.02573537 1.086609 1.086645 0.9351513 1.262639
#>     BM07 1.097759 0.02705855 1.097740 1.097778 0.9255588 1.301997
#>     BM08 1.093490 0.02731704 1.093471 1.093509 0.9205065 1.298980
#>     BM09 1.097057 0.02728325 1.097039 1.097076 0.9231711 1.303696
#>     BM10 1.102846 0.02636797 1.102828 1.102864 0.9399522 1.293969
#>     BM11 1.087660 0.02595190 1.087642 1.087677 0.9322930 1.268918
#>     BM12 1.097956 0.02736171 1.097937 1.097975 0.9236684 1.305130
#>     BM13 1.097559 0.02726128 1.097540 1.097578 0.9238865 1.303879
#>     BM14 1.094085 0.02713600 1.094067 1.094104 0.9219787 1.298319
#>     BM15 1.091205 0.02672594 1.091186 1.091223 0.9245129 1.287951
#>     BM16 1.093166 0.02728810 1.093147 1.093185 0.9206204 1.298051
#>     BM17 1.091134 0.02690460 1.091116 1.091153 0.9232650 1.289525
#>     BM18 1.101676 0.02663472 1.101658 1.101695 0.9352202 1.297759
#>     BM19 1.101516 0.02657692 1.101498 1.101534 0.9353070 1.297261
#>     BM20 1.109372 0.02394081 1.109355 1.109389 0.9860657 1.248097
#>     BM21 1.108795 0.02412337 1.108778 1.108812 0.9810365 1.253192
#>     BM22 1.091877 0.02691279 1.091858 1.091895 0.9231117 1.291496
#>     BM23 1.101041 0.02666699 1.101022 1.101059 0.9336578 1.298432
#>     BM24 1.094289 0.02737374 1.094270 1.094307 0.9204231 1.300997
#>  heterogeneity_variance       QE        QEp success error estimate_change
#>             0.006320850 35.97212 0.03057576    TRUE  <NA>   -0.0052058618
#>             0.005864195 35.04740 0.03830553    TRUE  <NA>   -0.0065734274
#>             0.006462614 36.45936 0.02709520    TRUE  <NA>   -0.0041016795
#>             0.007107127 37.47668 0.02095794    TRUE  <NA>    0.0001230324
#>             0.007015988 37.44715 0.02111657    TRUE  <NA>   -0.0005753113
#>             0.005204638 33.56675 0.05433410    TRUE  <NA>   -0.0084170853
#>             0.006846737 37.32105 0.02180625    TRUE  <NA>    0.0017752667
#>             0.006973427 37.19883 0.02249425    TRUE  <NA>   -0.0021212600
#>             0.007008180 37.42583 0.02123177    TRUE  <NA>    0.0011359233
#>             0.005953906 35.18667 0.03703967    TRUE  <NA>    0.0063985201
#>             0.005511142 34.22825 0.04656036    TRUE  <NA>   -0.0074671323
#>             0.007029085 37.30853 0.02187584    TRUE  <NA>    0.0019546482
#>             0.006980889 37.36588 0.02155870    TRUE  <NA>    0.0015931927
#>             0.006889116 37.31189 0.02185714    TRUE  <NA>   -0.0015767474
#>             0.006439218 36.40549 0.02746152    TRUE  <NA>   -0.0042133196
#>             0.006937412 37.12371 0.02292688    TRUE  <NA>   -0.0024173901
#>             0.006541070 36.43959 0.02722912    TRUE  <NA>   -0.0042778622
#>             0.006275548 35.91649 0.03099782    TRUE  <NA>    0.0053373184
#>             0.006258331 35.93542 0.03085360    TRUE  <NA>    0.0051919434
#>             0.003040836 29.45900 0.13221693    TRUE  <NA>    0.0122984848
#>             0.003319338 29.49158 0.13135418    TRUE  <NA>    0.0117785773
#>             0.006614442 36.71017 0.02544691    TRUE  <NA>   -0.0035974233
#>             0.006367544 36.19428 0.02894167    TRUE  <NA>    0.0047604664
#>             0.007044095 37.34628 0.02166661    TRUE  <NA>   -0.0013910973
#>  heterogeneity_change
#>          6.221489e-05
#>         -3.944398e-04
#>          2.039795e-04
#>          8.484926e-04
#>          7.573533e-04
#>         -1.053997e-03
#>          5.881023e-04
#>          7.147927e-04
#>          7.495452e-04
#>         -3.047285e-04
#>         -7.474929e-04
#>          7.704503e-04
#>          7.222547e-04
#>          6.304812e-04
#>          1.805829e-04
#>          6.787774e-04
#>          2.824352e-04
#>          1.691372e-05
#>         -3.032592e-07
#>         -3.217798e-03
#>         -2.939296e-03
#>          3.558077e-04
#>          1.089093e-04
#>          7.854609e-04
# Example 2: explicit effect-level sensitivity as a diagnostic.
apm_leave_one_out(apm_fit(agri_effects_benchmark), unit="effect")
#> <apm_sensitivity>
#>  omitted estimate         se ci_lower ci_upper  pi_lower pi_upper
#>        1 1.090122 0.02672415 1.090104 1.090140 0.9248696 1.284901
#>        2 1.088632 0.02629307 1.088614 1.088650 0.9288865 1.275850
#>        3 1.091326 0.02674515 1.091308 1.091345 0.9243547 1.288459
#>        4 1.095947 0.02742421 1.095928 1.095966 0.9211461 1.303918
#>        5 1.095182 0.02728954 1.095163 1.095200 0.9215091 1.301585
#>        6 1.086627 0.02573537 1.086609 1.086645 0.9351513 1.262639
#>        7 1.097759 0.02705855 1.097740 1.097778 0.9255588 1.301997
#>        8 1.093490 0.02731704 1.093471 1.093509 0.9205065 1.298980
#>        9 1.097057 0.02728325 1.097039 1.097076 0.9231711 1.303696
#>       10 1.102846 0.02636797 1.102828 1.102864 0.9399522 1.293969
#>       11 1.087660 0.02595190 1.087642 1.087677 0.9322930 1.268918
#>       12 1.097956 0.02736171 1.097937 1.097975 0.9236684 1.305130
#>       13 1.097559 0.02726128 1.097540 1.097578 0.9238865 1.303879
#>       14 1.094085 0.02713600 1.094067 1.094104 0.9219787 1.298319
#>       15 1.091205 0.02672594 1.091186 1.091223 0.9245129 1.287951
#>       16 1.093166 0.02728810 1.093147 1.093185 0.9206204 1.298051
#>       17 1.091134 0.02690460 1.091116 1.091153 0.9232650 1.289525
#>       18 1.101676 0.02663472 1.101658 1.101695 0.9352202 1.297759
#>       19 1.101516 0.02657692 1.101498 1.101534 0.9353070 1.297261
#>       20 1.109372 0.02394081 1.109355 1.109389 0.9860657 1.248097
#>       21 1.108795 0.02412337 1.108778 1.108812 0.9810365 1.253192
#>       22 1.091877 0.02691279 1.091858 1.091895 0.9231117 1.291496
#>       23 1.101041 0.02666699 1.101022 1.101059 0.9336578 1.298432
#>       24 1.094289 0.02737374 1.094270 1.094307 0.9204231 1.300997
#>  heterogeneity_variance       QE        QEp success error estimate_change
#>             0.006320850 35.97212 0.03057576    TRUE  <NA>   -0.0052058618
#>             0.005864195 35.04740 0.03830553    TRUE  <NA>   -0.0065734274
#>             0.006462614 36.45936 0.02709520    TRUE  <NA>   -0.0041016795
#>             0.007107127 37.47668 0.02095794    TRUE  <NA>    0.0001230324
#>             0.007015988 37.44715 0.02111657    TRUE  <NA>   -0.0005753113
#>             0.005204638 33.56675 0.05433410    TRUE  <NA>   -0.0084170853
#>             0.006846737 37.32105 0.02180625    TRUE  <NA>    0.0017752667
#>             0.006973427 37.19883 0.02249425    TRUE  <NA>   -0.0021212600
#>             0.007008180 37.42583 0.02123177    TRUE  <NA>    0.0011359233
#>             0.005953906 35.18667 0.03703967    TRUE  <NA>    0.0063985201
#>             0.005511142 34.22825 0.04656036    TRUE  <NA>   -0.0074671323
#>             0.007029085 37.30853 0.02187584    TRUE  <NA>    0.0019546482
#>             0.006980889 37.36588 0.02155870    TRUE  <NA>    0.0015931927
#>             0.006889116 37.31189 0.02185714    TRUE  <NA>   -0.0015767474
#>             0.006439218 36.40549 0.02746152    TRUE  <NA>   -0.0042133196
#>             0.006937412 37.12371 0.02292688    TRUE  <NA>   -0.0024173901
#>             0.006541070 36.43959 0.02722912    TRUE  <NA>   -0.0042778622
#>             0.006275548 35.91649 0.03099782    TRUE  <NA>    0.0053373184
#>             0.006258331 35.93542 0.03085360    TRUE  <NA>    0.0051919434
#>             0.003040836 29.45900 0.13221693    TRUE  <NA>    0.0122984848
#>             0.003319338 29.49158 0.13135418    TRUE  <NA>    0.0117785773
#>             0.006614442 36.71017 0.02544691    TRUE  <NA>   -0.0035974233
#>             0.006367544 36.19428 0.02894167    TRUE  <NA>    0.0047604664
#>             0.007044095 37.34628 0.02166661    TRUE  <NA>   -0.0013910973
#>  heterogeneity_change
#>          6.221489e-05
#>         -3.944398e-04
#>          2.039795e-04
#>          8.484926e-04
#>          7.573533e-04
#>         -1.053997e-03
#>          5.881023e-04
#>          7.147927e-04
#>          7.495452e-04
#>         -3.047285e-04
#>         -7.474929e-04
#>          7.704503e-04
#>          7.222547e-04
#>          6.304812e-04
#>          1.805829e-04
#>          6.787774e-04
#>          2.824352e-04
#>          1.691372e-05
#>         -3.032592e-07
#>         -3.217798e-03
#>         -2.939296e-03
#>          3.558077e-04
#>          1.089093e-04
#>          7.854609e-04
# Example 3: report results on the percent-change scale.
apm_leave_one_out(apm_fit(agri_effects_benchmark), transform="percent")
#> <apm_sensitivity>
#>  omitted  estimate         se  ci_lower  ci_upper  pi_lower pi_upper
#>     BM01  9.012202 0.02672415  9.010376  9.014028 -7.513039 28.49011
#>     BM02  8.863222 0.02629307  8.861429  8.865016 -7.111348 27.58503
#>     BM03  9.132638 0.02674515  9.130809  9.134467 -7.564530 28.84591
#>     BM04  9.594667 0.02742421  9.592784  9.596550 -7.885389 30.39181
#>     BM05  9.518159 0.02728954  9.516286  9.520032 -7.849087 30.15853
#>     BM06  8.662701 0.02573537  8.660948  8.664453 -6.484874 26.26388
#>     BM07  9.775893 0.02705855  9.774031  9.777754 -7.444119 30.19969
#>     BM08  9.348980 0.02731704  9.347108  9.350852 -7.949348 29.89804
#>     BM09  9.705731 0.02728325  9.703855  9.707606 -7.682888 30.36963
#>     BM10 10.284589 0.02636797 10.282767 10.286412 -6.004785 29.39691
#>     BM11  8.765974 0.02595190  8.764205  8.767743 -6.770697 26.89183
#>     BM12  9.795586 0.02736171  9.793704  9.797469 -7.633156 30.51297
#>     BM13  9.755907 0.02726128  9.754032  9.757782 -7.611347 30.38786
#>     BM14  9.408538 0.02713600  9.406678  9.410399 -7.802126 29.83193
#>     BM15  9.120455 0.02672594  9.118627  9.122282 -7.548712 28.79511
#>     BM16  9.316604 0.02728810  9.314734  9.318473 -7.937959 29.80507
#>     BM17  9.113412 0.02690460  9.111573  9.115252 -7.673502 28.95254
#>     BM18 10.167617 0.02663472 10.165779 10.169456 -6.477977 29.77589
#>     BM19 10.151603 0.02657692 10.149768 10.153437 -6.469303 29.72613
#>     BM20 10.937188 0.02394081 10.935524 10.938852 -1.393429 24.80973
#>     BM21 10.879526 0.02412337 10.877850 10.881202 -1.896354 25.31919
#>     BM22  9.187682 0.02691279  9.185841  9.189524 -7.688835 29.14960
#>     BM23 10.104085 0.02666699 10.102245 10.105925 -6.634222 29.84318
#>     BM24  9.428852 0.02737374  9.426975  9.430729 -7.957695 30.09967
#>  heterogeneity_variance       QE        QEp success error estimate_change
#>             0.006320850 35.97212 0.03057576    TRUE  <NA>   -0.0052058618
#>             0.005864195 35.04740 0.03830553    TRUE  <NA>   -0.0065734274
#>             0.006462614 36.45936 0.02709520    TRUE  <NA>   -0.0041016795
#>             0.007107127 37.47668 0.02095794    TRUE  <NA>    0.0001230324
#>             0.007015988 37.44715 0.02111657    TRUE  <NA>   -0.0005753113
#>             0.005204638 33.56675 0.05433410    TRUE  <NA>   -0.0084170853
#>             0.006846737 37.32105 0.02180625    TRUE  <NA>    0.0017752667
#>             0.006973427 37.19883 0.02249425    TRUE  <NA>   -0.0021212600
#>             0.007008180 37.42583 0.02123177    TRUE  <NA>    0.0011359233
#>             0.005953906 35.18667 0.03703967    TRUE  <NA>    0.0063985201
#>             0.005511142 34.22825 0.04656036    TRUE  <NA>   -0.0074671323
#>             0.007029085 37.30853 0.02187584    TRUE  <NA>    0.0019546482
#>             0.006980889 37.36588 0.02155870    TRUE  <NA>    0.0015931927
#>             0.006889116 37.31189 0.02185714    TRUE  <NA>   -0.0015767474
#>             0.006439218 36.40549 0.02746152    TRUE  <NA>   -0.0042133196
#>             0.006937412 37.12371 0.02292688    TRUE  <NA>   -0.0024173901
#>             0.006541070 36.43959 0.02722912    TRUE  <NA>   -0.0042778622
#>             0.006275548 35.91649 0.03099782    TRUE  <NA>    0.0053373184
#>             0.006258331 35.93542 0.03085360    TRUE  <NA>    0.0051919434
#>             0.003040836 29.45900 0.13221693    TRUE  <NA>    0.0122984848
#>             0.003319338 29.49158 0.13135418    TRUE  <NA>    0.0117785773
#>             0.006614442 36.71017 0.02544691    TRUE  <NA>   -0.0035974233
#>             0.006367544 36.19428 0.02894167    TRUE  <NA>    0.0047604664
#>             0.007044095 37.34628 0.02166661    TRUE  <NA>   -0.0013910973
#>  heterogeneity_change
#>          6.221489e-05
#>         -3.944398e-04
#>          2.039795e-04
#>          8.484926e-04
#>          7.573533e-04
#>         -1.053997e-03
#>          5.881023e-04
#>          7.147927e-04
#>          7.495452e-04
#>         -3.047285e-04
#>         -7.474929e-04
#>          7.704503e-04
#>          7.222547e-04
#>          6.304812e-04
#>          1.805829e-04
#>          6.787774e-04
#>          2.824352e-04
#>          1.691372e-05
#>         -3.032592e-07
#>         -3.217798e-03
#>         -2.939296e-03
#>          3.558077e-04
#>          1.089093e-04
#>          7.854609e-04